Handle that the static dir tiny_mce could be composed from multiple sources
With the app staticfiles it is possible that the static dir tiny_mce is
composed of multiple sources. As example I have a project which uses a
helper app which contains a TinyMCE skin under
tiny_mce/themes/advanced/skins/skinX in it's static dir. The helper app
is listed before tinymce in INSTALLED_APPS and therefore the old code
determined helper_app/static/tiny_mce as base_path in get_file_contents
and this was not right because most of the TinyMCE files were under
tinymce/static/tiny_mce. By using the finder for every individual file
this problem is solved.